rationalized and finished JSON options
